OptimizationResult becomes OptimizationState (#297)
* move optresult in alg

* rename OptimizationResult -> OptimizationState
