+++ src/Algorithm/Algorithm.jl
@@ -4,7 +4,7 @@ import DataStructures
 import MathOptInterface
 import TimerOutputs
 
-import ..Coluna
+using ..Coluna
 using ..Containers
 using ..MathProg
 
@@ -24,11 +24,19 @@ const MOI = MathOptInterface
 
 import Base: push!
 
-# Abstract algorithm
-include("interface.jl")
+# Import to extend methods to OptimizationState
+import .MathProg: getfeasibilitystatus, getterminationstatus, setfeasibilitystatus!,
+    setterminationstatus!, isfeasible, get_ip_primal_bound, get_ip_dual_bound, 
+    get_lp_primal_bound, get_lp_dual_bound, update_ip_primal_bound!, update_ip_dual_bound!, 
+    update_lp_primal_bound!, update_lp_dual_bound!, set_ip_primal_bound!, 
+    set_ip_dual_bound!, set_lp_primal_bound!, set_lp_dual_bound!, ip_gap
 
-# Abstract record
-include("record.jl")
+# Utilities to build algorithms
+include("utilities/optimizationstate.jl")
+include("utilities/record.jl")
+
+# Abstract algorithm (interface should be moved in Containers)
+include("interface.jl")
 
 # Here include slave algorithms used by conquer algorithms
 include("solveipform.jl")
@@ -52,11 +60,21 @@ include("branching/branchingalgo.jl")
 
 include("treesearch.jl")
 
-# Types
+# Algorithm should export only methods usefull to define & parametrize algorithms, and 
+# data structures from utilities.
+# Other Coluna's submodules should be independent to Algorithm
+
+# Utilities
+export OptimizationState, getterminationstatus, getfeasibilitystatus, setterminationstatus!,
+    setfeasibilitystatus!, isfeasible, nb_ip_primal_sols, nb_lp_primal_sols, nb_lp_dual_sols,
+    get_ip_primal_sols, get_lp_primal_sols, get_lp_dual_sols, get_best_ip_primal_sol,
+    get_best_lp_primal_sol, get_best_lp_dual_sol, update_ip_primal_sol!, 
+    update_lp_primal_sol!, update_lp_dual_sol!, add_ip_primal_sol!, add_lp_primal_sol!,
+    add_lp_dual_sol!, set_ip_primal_sol!, set_lp_primal_sol!, set_lp_dual_sol!
+
+# Algorithm's types
 export AbstractOptimizationAlgorithm, TreeSearchAlgorithm, ColGenConquer, ColumnGeneration, 
        BendersConquer, BendersCutGeneration, SolveIpForm, SolveLpForm, ExactBranchingPhase, 
        OnlyRestrictedMasterBranchingPhase
-
-export getinputresult
 
 end
