+++ src/MOIwrapper.jl
@@ -15,12 +15,12 @@ const SupportedConstrSets = Union{MOI.EqualTo{Float64},
 
 
 mutable struct Optimizer <: MOI.AbstractOptimizer
-    inner::Union{Nothing, Problem}
+    inner::Union{Nothing,Problem}
     moi_index_to_coluna_uid::MOIU.IndexMap
     params::Params
     annotations::Annotations
     varmap::Dict{MOI.VariableIndex,VarId} # For the user to get VariablePrimal
-    result::Union{Nothing,OptimizationResult}
+    result::Union{Nothing,OptimizationState}
 end
 
 setinnerprob!(o::Optimizer, prob::Problem) = o.inner = prob
